backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "examen S2G"
authors = ["maxiR143 <maximilianovl98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.11"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = "^1.7.4"
peewee = "^3.17.9"
pydantic-settings = "^2.8.1"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
